requirements.txt
langchain
langchain-community
langchain-openai
huggingface-hub
transformers
tensorflow
torch
langchain-core
streamlit
python-dotenv
langserve
fastapi
uvicorn
sse_starlette
pydantic
langchain-ollama
requests
pydantic<2
bs4
pypdf
chromadb
faiss-cpu